Bring back IcePy.compile for slice2py script
