+++ python/modules/IcePy/Slice.h
@@ -6,5 +6,6 @@
 #include "Config.h"
 
 extern "C" PyObject* IcePy_loadSlice(PyObject*, PyObject*);
+extern "C" PyObject* IcePy_compileSlice(PyObject*, PyObject*);
 
 #endif
